% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-page.R
\name{crunchPage}
\alias{crunchPage}
\alias{crunchFluidPage}
\alias{crunchFillPage}
\alias{crunchNavbarPage}
\title{Build a Crunchy UI}
\usage{
crunchPage(...)

crunchFluidPage(...)

crunchFillPage(...)

crunchNavbarPage(...)
}
\arguments{
\item{...}{arguments passed to \code{fluidPage}, \code{fillPage} or \code{navbarPage}}
}
\value{
The result of \code{fluidPage}, \code{fillPage} or \code{navbarPage}
}
\description{
Shiny provides page functions for defining UI layout. These
functions wraps those and includes additional assets needed to match
Crunch UI style and keep your users authenticated. When building shiny apps
with Crunch datasets, use these instead of \code{\link[shiny:fluidPage]{shiny::fluidPage()}},
\code{\link[shiny:fillPage]{shiny::fillPage()}} or \code{\link[shiny:navbarPage]{shiny::navbarPage()}}.
}
\details{
These are no longer necessary. Just use the \code{shiny} ones and it just works.
These functions are left here for backwards compatibility.
}
\examples{
\dontrun{
crunchPage(
    fluidRow(
        column(6,
            selectInput("filter",
                label="Filter",
                choices=filterList,
                selected="All"),
            br(),
            plotOutput("funnel1", height="300"),
        ),
        column(6,
            selectInput("brand",
                label="Competitor",
                choices=brands,
                selected="Nike"),
            br(),
            plotOutput("funnel2", height="300"),
        )
    )
)
}
}
