% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{extract_subdomain}
\alias{extract_subdomain}
\alias{extract_domain}
\title{Extract components of URLs}
\usage{
extract_subdomain(url_hostname)

extract_domain(url_hostname)
}
\arguments{
\item{url_hostname}{The url hostname like \code{yougov.crunch.io}}
}
\description{
Extract components of URLs
}
