% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{launch_demo}
\alias{launch_demo}
\title{Run demo application}
\usage{
launch_demo(which = "demo_auth")
}
\arguments{
\item{which}{demo to run (currently only 'demo_auth' is supported)}
}
\description{
Run demo application
}
