% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{listDatasetGadget}
\alias{listDatasetGadget}
\title{Open dataset selector}
\usage{
listDatasetGadget(kind = c("active", "all", "archived"), refresh = FALSE)
}
\value{
A `loadDataset()` call is pasted into your RStudio session`
}
\description{
Open dataset selector
}
