% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchy_files.R
\name{open-netcdf-4}
\alias{open-netcdf-4}
\alias{open_oc_file_ncdf4}
\alias{open-netcdf-4}
\alias{open_oc_file_RNetCDF}
\title{Open files}
\usage{
open_oc_file_ncdf4()

open_oc_file_RNetCDF()
}
\description{
Open files

Title
}
