% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderPlot}
\alias{renderPlot}
\title{Wrapper around shiny's renderPlot for better error handling}
\usage{
renderPlot(expr, ...)
}
\arguments{
\item{expr}{expression that renders a plot}

\item{...}{additional arguments passed to \code{\link[shiny]{renderPlot}}}
}
\value{
Same as \code{renderPlot}, but if \code{expr} errors when evaluated,
instead of printing a red R error message in your app, you get a more
user-friendly placeholder.
}
\description{
Wrapper around shiny's renderPlot for better error handling
}

