% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-user.R
\name{shinyUser}
\alias{shinyUser}
\title{Return Crunch User Information}
\usage{
shinyUser()
}
\value{
A user record if the user is logged in, otherwise a character vector or error
}
\description{
This function returns information about the current user of the shiny app. This
is useful if you want to change the behavior of the app depending on who is
viewing the app. You can access elements of this record with crunch functions
like \code{name()} or \code{email().}
}
