% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{with_temp_auth}
\alias{with_temp_auth}
\title{Use Crunch API with the current user's authentication}
\usage{
with_temp_auth(auth, expr)
}
\arguments{
\item{auth}{An authentication object from the \code{authenticationServer} module}

\item{expr}{An expression that will run with authentication}
}
\description{
Because rcrunch is not designed for an environment where it has to make API
requests on behalf of multiple users, like a shiny server, we need to be
careful that we always set the authentication for the correct user.
}
\examples{
\dontrun{
 auth <- authenticationServer('auth')
 ds_rx <- reactive({with_temp_auth(auth, {
     loadDataset(dataset_id(), project = NULL)
 })})
}
}
