% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authenticationUI}
\alias{authenticationUI}
\alias{authenticationServer}
\title{Cookie-based authentication}
\usage{
authenticationUI(id, req, invisible = FALSE)

authenticationServer(id, restrict_to = list(), dev_mode = list())
}
\arguments{
\item{id}{A string that identifies the shiny module between UI and Server objects
(typically just "auth")}

\item{req}{The request object, received by using a function for your shiny's UI,
defaults to an empty list, which means no restriction. See details.}

\item{invisible}{Whether to hide the authentication UI (the user's email as a clickable link)
defaults to \code{FALSE}}

\item{restrict_to}{\code{list} describing who should have access to the shiny application, see details}

\item{dev_mode}{An optional list of overrides for authentication components (\code{email}, \code{subdomain}, \code{domain},
\code{api_key}, and \code{access}) for use during development.}
}
\description{
Cookie-based authentication
}
\details{
The \code{restrict_to} argument defaults to an empty list, which allows unrestricted access.
To restrict it, add one or more named elements:
\itemize{
\item \code{email_in}: Character vector of exact email addresses
\item \code{email_like}: String pattern to test user's email address against (passed to \code{\link[=grepl]{grepl()}})
\item \code{can_load_dataset}: String with Crunch dataset ID that the user must have access to
}

If multiple elements are supplied, the user will gain access if \strong{at least one} of them works.
}
\examples{
# UI should be a function, passing its argument to authenticationUI
ui <- function(req){
     navbarPage(
         title = 'Some App',
         navbarMenu('Some Menu',tabPanel('Some panel', p('Hello'))),
         bslib::nav_spacer(),
         bslib::nav_item(
             authenticationUI('auth', req)
         )
     )
}
# server function
server <- function(input, output, session){
    auth <- authenticationServer(
        'auth',
        restrict_to = list(
             email_in         = 'hinko@client.com',
             email_like       = '*@crunch.io',
             can_load_dataset = 'wftcfkglaelnfgehcgqsfvnkuijyzwfb'
        )
    )
    foo <- fooServer(
         'foo',
         access    = auth$access,
         subdomain = auth$subdomain,
         domain    = auth$domain,
         user      = auth$email,
         api_key   = auth$api_key
    )
}
}
