% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-page.R
\name{crunchPage}
\alias{crunchPage}
\alias{crunchFluidPage}
\alias{crunchFillPage}
\alias{crunchNavbarPage}
\title{Build a Crunchy UI}
\usage{
crunchPage(...)

crunchFluidPage(...)

crunchFillPage(...)

crunchNavbarPage(...)
}
\arguments{
\item{...}{arguments passed to \code{fluidPage}, \code{fillPage} or \code{navbarPage}}
}
\value{
The result of \code{fluidPage}, \code{fillPage} or \code{navbarPage}
}
\description{
These are no longer necessary. Just use the \code{shiny} ones and it just works.
These functions are left here for backwards compatibility.
}
\examples{
\dontrun{
crunchPage(
    fluidRow(
        column(6,
            selectInput("filter",
                label="Filter",
                choices=filterList,
                selected="All"),
            br(),
            plotOutput("funnel1", height="300"),
        ),
        column(6,
            selectInput("brand",
                label="Competitor",
                choices=brands,
                selected="Nike"),
            br(),
            plotOutput("funnel2", height="300"),
        )
    )
)
}
}
