% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-page.R
\name{crunchPage}
\alias{crunchFillPage}
\alias{crunchFluidPage}
\alias{crunchPage}
\title{Build a Crunchy UI}
\usage{
crunchPage(...)

crunchFluidPage(...)

crunchFillPage(...)
}
\arguments{
\item{...}{arguments passed to \code{fluidPage}}
}
\value{
The result of \code{fluidPage}
}
\description{
Shiny provides a \code{\link[shiny]{fluidPage}} for defining UI layout. This
function wraps that one and includes additional assets needed to match
Crunch UI style and keep your users authenticated. When building shiny apps
with Crunch datasets, use this instead of \code{fluidPage}.
}

