% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch-page.R
\name{crunchPage}
\alias{crunchPage}
\alias{crunchFluidPage}
\alias{crunchFillPage}
\alias{crunchNavbarPage}
\title{Build a Crunchy UI}
\usage{
crunchPage(...)

crunchFluidPage(...)

crunchFillPage(...)

crunchNavbarPage(title = NULL, ...)
}
\arguments{
\item{...}{arguments passed to `fluidPage`, `fillPage` or `navbarPage`}

\item{title}{the title to be displayed in the navigation bar. Defaults to NULL.}
}
\value{
The result of `fluidPage`, `fillPage` or `navbarPage`
}
\description{
Shiny provides page functions for defining UI layout. These
functions wraps those and includes additional assets needed to match
Crunch UI style and keep your users authenticated. When building shiny apps
with Crunch datasets, use these instead of [shiny::fluidPage()],
[shiny::fillPage()] or [shiny::navbarPage()].
}
\examples{
\dontrun{
crunchPage(
    fluidRow(
        column(6,
            selectInput("filter",
                label="Filter",
                choices=filterList,
                selected="All"),
            br(),
            plotOutput("funnel1", height="300"),
        ),
        column(6,
            selectInput("brand",
                label="Competitor",
                choices=brands,
                selected="Nike"),
            br(),
            plotOutput("funnel2", height="300"),
        )
    )
)
}
}
