% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{crunchy_page_navbar}
\alias{crunchy_page_navbar}
\alias{crunchy_page_navbar_server}
\title{Simple page_navbar replacement for crunchy authentication}
\usage{
crunchy_page_navbar(..., id = "crunchy-page-navbar")

crunchy_page_navbar_server(
  ...,
  id = "crunchy-page-navbar",
  restrict_to = list(),
  dev_mode = list(),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{..., id}{Passed to the named arguments of \code{\link[bslib:page_navbar]{bslib::page_navbar()}}.
Note that instead of putting the unnamed \code{bslib::nav_panel()} arguments
the ui, as you would in \code{bslib::page_navbar()} you instead put them in the server
inside \code{crunchy_page_navbar_server()}. The authentication is put after your nav items,
after a spacer.}

\item{restrict_to, dev_mode}{Passed to \code{\link[=authenticationServer]{authenticationServer()}}}

\item{session}{The shiny reactive domain (almost always can use the default)}
}
\value{
For \code{crunchy_page_navbar()} A function that can be used as a ui for \code{shiny::shinyApp()}
and for \code{crunchy_page_navbar_server()}, the auth module
}
\description{
This is a replacement for \code{\link[bslib:page_navbar]{bslib::page_navbar()}} that makes it easier to
put your shiny application's UI behind the crunchy authentication. To use
put the non-\code{bslib::nav_panel()} parts of a \code{page_navbar()} into \code{crunchy_page_navbar()}
and then put the nav_panels inside your shiny's server function inside \code{crunchy_page_navbar_server()}.
}
\details{
These navpanels will only show after the application has determined that the user has
access. A more customizable interface is possible with \code{\link[=authenticationUI]{authenticationUI()}} and
\code{\link[=authenticationServer]{authenticationServer()}} modules, but note that your app may start to load its UI
for anyone logged into crunch, whether or not they are supposed to have access to your app.
}
\examples{
\dontrun{
dev_mode <- list(
    api_key = crunch::envOrOption('crunch.api.key'),
    domain = extract_domain(gsub("https?://", "", crunch::envOrOption('crunch.api'))),
    subdomain = extract_subdomain(gsub("https?://", "", crunch::envOrOption('crunch.api')))
)
ui <- crunchy_page_navbar(title = "Some App")

# server function
server <- function(input, output, session){
    auth <- crunchy_page_navbar_server(
        dev_mode = dev_mode,
        restrict_to = list(email_like = "@"),
        bslib::nav_panel(
            title = "Welcome",
            uiOutput("welcome"),
            verbatimTextOutput("projects")
        )
    )

    output$welcome <- renderUI({tags$p("Welcome back ", auth$email(), "!")})
    output$projects <- renderText({
        with_temp_auth(auth, {
            paste0(capture.output(crunch::projects()), collapse = "\n")
        })
     })
}
shinyApp(ui, server)
}
}
