% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchy-server.R
\name{crunchyServer}
\alias{crunchyServer}
\title{A Shiny server with Crunch auth}
\usage{
crunchyServer(func, authz = getOption("crunchy.authorization"))
}
\arguments{
\item{func}{A `function (input, output, session)`, as you'd normally give to
`shinyServer()`. If the user is not authenticated or authorized, this
function will not be evaluated.}

\item{authz}{A `function (input, output, session)` or expression to evaluate
to determine if the current user is authorized to enter the app. Ideally,
this is cheap to execute because it will be called repeatedly.}
}
\value{
Invisibly, a Shiny server function with the auth logic wrapped around
`func`.
}
\description{
To make sure that users who access your shiny.crunch.io app are allowed to
access it, use `crunchyServer()` instead of [shiny::shinyServer()], and wrap
your UI content inside [crunchyBody()]. This will prevent anyone who is not
logged into Crunch in their browser from accessing your app.
}
\details{
To restrict access further to only certain Crunch users, you can set an
authorization method, either by passing a server function or expression to
the `authz` argument of this function, or by calling
[setCrunchyAuthorization()].
}
\seealso{
[crunchyBody()] for wrapping the UI contents, [crunchyPublicBody()]
for specifying an alternate UI for when the user is not authenticated, and
[crunchyUnauthorizedBody()] for giving an alternate UI for users who are
authenticated with Crunch but not authorized to view this app.
}
