% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget-list-datasets.R
\name{listDatasetGadget}
\alias{listDatasetGadget}
\title{Open dataset selector}
\usage{
listDatasetGadget(...)
}
\arguments{
\item{...}{Further arguments passed to \code{crunch::listDatasets()}}
}
\value{
A \code{loadDataset()} call is pasted into your RStudio session
}
\description{
Open dataset selector
}
