% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{listDatasetGadget}
\alias{listDatasetGadget}
\title{Open dataset selector}
\usage{
listDatasetGadget(...)
}
\arguments{
\item{...}{Further arguments passed to `crunch::listDatasets()`}
}
\value{
A `loadDataset()` call is pasted into your RStudio session`
}
\description{
Open dataset selector
}
\keyword{internal}
