% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-record.R
\name{shinyUser}
\alias{shinyUser}
\alias{checkAuthentication}
\title{Get Crunch user details and confirm authentication}
\usage{
shinyUser()

checkAuthentication()
}
\value{
A user record if the user is logged in, otherwise a character vector or error
}
\description{
`shinyUser()` returns information about the current user of the Shiny app. This
is useful if you want to change the behavior of the app depending on who is
viewing the app. You can access elements of this record with Crunch functions
like `name()` or `email().`
}
\details{
`checkAuthentication()` is an alias for `shinyUser()` for when you only care
about ensuring that the current user is authenticated with Crunch.
}
